fill area, consolidate hover state color
