+++ web/src/common/index.ts
@@ -17,6 +17,7 @@ export { default as QualitativeLegend } from "./QualitativeLegend.svelte";
 export { default as SequentialLegend } from "./SequentialLegend.svelte";
 export { default as StreetView } from "./StreetView.svelte";
 export { layerId } from "./zorder";
+export { Style } from "./colors";
 
 // TS fix for the imprecise geojson types
 export function gjPosition(pt: number[]): [number, number] {
